+++ test/integration/zz_generated.registered_testcases.go
@@ -22,6 +22,7 @@ func init() {
 		TestGatewayClassCreation,
 		TestGatewayClassUpdates,
 		TestGatewayConfigurationEssentials,
+		TestGatewayConfigurationServiceName,
 		TestGatewayDataPlaneNetworkPolicy,
 		TestGatewayEssentials,
 		TestGatewayMultiple,